requirements.txt
torch>=2.0.0
mmsegmentation>=1.0.0
mmcv>=2.0.0
mmdet>=3.0.0
rasterio
geopandas
shapely
opencv-python
pytorch-lightning
albumentations>=1.3.0
